api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry  xx

azure-storage-blob
azure-identity
azure-functions
azure-monitor-query
requests
